requirements.txt
langchain
pypdf
python-dotenv==1.0.0
streamlit==1.22.0
faiss-cpu==1.7.4
altair==4.0
qdrant-client
